+++ lib/Sema/TypeCheckDeclOverride.cpp
@@ -19,10 +19,11 @@
 #include "TypeCheckDecl.h"
 #include "TypeCheckEffects.h"
 #include "TypeCheckObjC.h"
-#include "TypeChecker.h"
 #include "TypeCheckUnsafe.h"
+#include "TypeChecker.h"
 #include "swift/AST/ASTVisitor.h"
-#include "swift/AST/Availability.h"
+#include "swift/AST/AvailabilityInference.h"
+#include "swift/AST/AvailabilityRange.h"
 #include "swift/AST/Decl.h"
 #include "swift/AST/GenericEnvironment.h"
 #include "swift/AST/GenericSignature.h"
